Linq

+++ LAB_2018PR603_2019AM604/Models/Nota.cs
@@ -9,8 +9,8 @@ namespace LAB_2018PR603_2019AM604.Models
         public int id { get; set; }
         public int inscripcionId { get; set; }
         public string evaluacion { get; set; }
-        public double nota { get; set; }
-        public double porcentaje { get; set; }
+        public decimal nota { get; set; }
+        public decimal porcentaje { get; set; }
         public DateTime fecha { get; set; }
     }
 }
